Refactor code

+++ CGCourse/CGCourse/CGCourse.cpp
@@ -20,7 +20,7 @@
 
 using namespace std;
 extern "C" {
-	_declspec(dllexport) DWORD NvOptimusEnablement = 0x00000001;
+	_declspec(dllexport) unsigned long NvOptimusEnablement = 0x00000001;
 }
 Game::World* world = nullptr;
